requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
sqlalchemy[asyncio]>=2.0.0
asyncpg
alembic
pydantic>=2.0.0
pydantic-settings
python-jose[cryptography]
python-dotenv
httpx
boto3
python-multipart
